Aws.DynamoDb.Commands.Query

Description

Implementation of Amazon DynamoDb Query command.

See: http://docs.aws.amazon.com/amazondynamodb/latest/APIReference/API_Query.html

Documentation

data Query

A Query command that uses primary keys for an expedient scan.

Constructors

Query

Fields

qTableName :: Text

Required.

qKeyConditions :: Slice

Required. Hash or hash-range main condition.

qFilter :: Conditions

Whether to filter results before returning to client

qStartKey :: Maybe [Attribute]

Exclusive start key to resume a previous query.

qLimit :: Maybe Int

Whether to limit result set size

qForwardScan :: Bool

Set to False for descending results

qSelect :: QuerySelect

What to return from Query

qRetCons :: ReturnConsumption

qIndex :: Maybe Text

Whether to use a secondary/global index

qConsistent :: Bool

data Slice

Slice is the primary constraint in a Query command, per AWS requirements.

All Query commands must specify a hash attribute via DEq and optionally provide a secondary range attribute.

Constructors

Slice

Fields

sliceHash :: Attribute

Hash value of the primary key or index being used

sliceCond :: Maybe Condition

An optional condition specified on the range component, if present, of the primary key or index being used.

query

Arguments

:: Text
Table name

-> Slice
Primary key slice for query

-> Query

Construct a minimal Query request.

data QueryResponse

Response to a Query query.

Constructors

QueryResponse

Fields

qrItems :: Vector Item

qrLastKey :: Maybe [Attribute]

qrCount :: Int

qrScanned :: Int

qrConsumed :: Maybe ConsumedCapacity