Aws.DynamoDb.Commands.Table

Commands

data CreateTable

Constructors

CreateTable

Fields

createTableName :: Text

createAttributeDefinitions :: [AttributeDefinition]

only attributes appearing in a key must be listed here

createKeySchema :: KeySchema

createProvisionedThroughput :: ProvisionedThroughput

createLocalSecondaryIndexes :: [LocalSecondaryIndex]

at most 5 local secondary indices are allowed

createGlobalSecondaryIndexes :: [GlobalSecondaryIndex]

createTable

Arguments

:: Text
Table name

-> [AttributeDefinition]

-> KeySchema

-> ProvisionedThroughput

-> CreateTable

newtype CreateTableResult

Constructors

CreateTableResult

Fields

ctStatus :: TableDescription

data DescribeTable

Constructors

DescribeTable

Fields

dTableName :: Text

newtype DescribeTableResult

Constructors

DescribeTableResult

Fields

dtStatus :: TableDescription

data UpdateTable

Constructors

UpdateTable

Fields

updateTableName :: Text

updateProvisionedThroughput :: ProvisionedThroughput

updateGlobalSecondaryIndexUpdates :: [GlobalSecondaryIndexUpdate]

newtype UpdateTableResult

Constructors

UpdateTableResult

Fields

uStatus :: TableDescription

data DeleteTable

Constructors

DeleteTable

Fields

deleteTableName :: Text

newtype DeleteTableResult

Constructors

DeleteTableResult

Fields

dStatus :: TableDescription

data ListTables

TODO: currently this does not support restarting a cutoff query because of size.

Constructors

ListTables

newtype ListTablesResult

Constructors

ListTablesResult

Fields

tableNames :: [Text]

Data passed in the commands

data AttributeType

The type of a key attribute that appears in the table key or as a key in one of the indices.

Constructors

AttrString

AttrNumber

AttrBinary

data AttributeDefinition

A key attribute that appears in the table key or as a key in one of the indices.

Constructors

AttributeDefinition

Fields

attributeName :: Text

attributeType :: AttributeType

data KeySchema

The key schema can either be a hash of a single attribute name or a hash attribute name and a range attribute name.

Constructors

HashOnly Text

HashAndRange Text Text

data Projection

This determines which attributes are projected into a secondary index.

Constructors

ProjectKeysOnly

ProjectAll

ProjectInclude [Text]

data LocalSecondaryIndex

Describes a single local secondary index. The KeySchema MUST share the same hash key attribute as the parent table, only the range key can differ.

Constructors

LocalSecondaryIndex

Fields

localIndexName :: Text

localKeySchema :: KeySchema

localProjection :: Projection

data LocalSecondaryIndexStatus

This is returned by AWS to describe the local secondary index.

Constructors

LocalSecondaryIndexStatus

Fields

locStatusIndexName :: Text

locStatusIndexSizeBytes :: Integer

locStatusItemCount :: Integer

locStatusKeySchema :: KeySchema

locStatusProjection :: Projection

data ProvisionedThroughput

The target provisioned throughput you are requesting for the table or global secondary index.

Constructors

ProvisionedThroughput

Fields

readCapacityUnits :: Int

writeCapacityUnits :: Int

data ProvisionedThroughputStatus

This is returned by AWS as the status of the throughput for a table or global secondary index.

Constructors

ProvisionedThroughputStatus

Fields

statusLastDecreaseDateTime :: UTCTime

statusLastIncreaseDateTime :: UTCTime

statusNumberOfDecreasesToday :: Int

statusReadCapacityUnits :: Int

statusWriteCapacityUnits :: Int

data GlobalSecondaryIndex

Describes a global secondary index.

Constructors

GlobalSecondaryIndex

Fields

globalIndexName :: Text

globalKeySchema :: KeySchema

globalProjection :: Projection

globalProvisionedThroughput :: ProvisionedThroughput

data GlobalSecondaryIndexStatus

This is returned by AWS to describe the status of a global secondary index.

Constructors

GlobalSecondaryIndexStatus

Fields

gStatusIndexName :: Text

gStatusIndexSizeBytes :: Integer

gStatusIndexStatus :: Text

gStatusItemCount :: Integer

gStatusKeySchema :: KeySchema

gStatusProjection :: Projection

gStatusProvisionedThroughput :: ProvisionedThroughputStatus

data GlobalSecondaryIndexUpdate

This is used to request a change in the provisioned throughput of a global secondary index as part of an UpdateTable operation.

Constructors

GlobalSecondaryIndexUpdate

Fields

gUpdateIndexName :: Text

gUpdateProvisionedThroughput :: ProvisionedThroughput

data TableDescription

This describes the table and is the return value from AWS for all the table-related commands.

Constructors

TableDescription

Fields

rTableName :: Text

rTableSizeBytes :: Integer

rTableStatus :: Text

one of CREATING, UPDATING, DELETING, ACTIVE

rCreationDateTime :: Maybe UTCTime

rItemCount :: Integer

rAttributeDefinitions :: [AttributeDefinition]

rKeySchema :: Maybe KeySchema

rProvisionedThroughput :: ProvisionedThroughputStatus

rLocalSecondaryIndexes :: [LocalSecondaryIndexStatus]

rGlobalSecondaryIndexes :: [GlobalSecondaryIndexStatus]