Aws.DynamoDb.Core

Description

Shared types and utilities for DyanmoDb functionality.

Configuration and Regions

data Region

Constructors

Region

Fields

rUri :: ByteString

rName :: ByteString

ddbLocal :: Region

DynamoDb local connection (for development)

ddbUsEast1 :: Region

ddbUsWest1 :: Region

ddbUsWest2 :: Region

ddbEuWest1 :: Region

ddbEuWest2 :: Region

ddbEuCentral1 :: Region

ddbApNe1 :: Region

ddbApSe1 :: Region

ddbApSe2 :: Region

ddbSaEast1 :: Region

data DdbConfiguration qt

Constructors

DdbConfiguration

Fields

ddbcRegion :: Region

The regional endpoint. Ex: ddbUsEast

ddbcProtocol :: Protocol

HTTP or HTTPS

ddbcPort :: Maybe Int

Port override (mostly for local dev connection)

DynamoDB values

data DValue

Value types natively recognized by DynamoDb. We pretty much exactly reflect the AWS API onto Haskell types.

Constructors

DNull

DNum Scientific

DString Text

DBinary ByteString
Binary data will automatically be base64 marshalled.

DNumSet (Set Scientific)

DStringSet (Set Text)

DBinSet (Set ByteString)
Binary data will automatically be base64 marshalled.

DBool Bool

DBoolSet (Set Bool)
Composite data

DList (Vector DValue)

DMap (Map Text DValue)

Converting to/from DValue

class DynData (DynRep a) => DynVal a where

Class of Haskell types that can be represented as DynamoDb values.

This is the conversion layer; instantiate this class for your own types and then use the toValue and fromValue combinators to convert in application code.

Each Haskell type instantiated with this class will map to a DynamoDb-supported type that most naturally represents it.

Associated Types

type DynRep a

Which of the DynData instances does this data type directly map to?

Methods

toRep :: a -> DynRep a

Convert to representation

fromRep :: DynRep a -> Maybe a

Convert from representation

toValue :: DynVal a => a -> DValue

Encode a Haskell value.

fromValue :: DynVal a => DValue -> Maybe a

Decode a Haskell value.

newtype Bin a

Type wrapper for binary data to be written to DynamoDB. Wrap any Serialize instance in there and DynVal will know how to automatically handle conversions in binary form.

Constructors

Bin

Fields

getBin :: a

newtype OldBool

Constructors

OldBool Bool

Defining new DynVal instances

class Ord a => DynData a where

An internally used closed typeclass for values that have direct DynamoDb representations. Based on AWS API, this is basically numbers, strings and binary blobs.

This is here so that any DynVal haskell value can automatically be lifted to a list or a Set without any instance code duplication.

Do not try to create your own instances.

Methods

fromData :: a -> DValue

toData :: DValue -> Maybe a

newtype DynBinary

Binary values stored in DynamoDb. Only used in defining new DynVal instances.

Constructors

DynBinary

Fields

unDynBinary :: ByteString

newtype DynNumber

Numeric values stored in DynamoDb. Only used in defining new DynVal instances.

Constructors

DynNumber

Fields

unDynNumber :: Scientific

newtype DynString

String values stored in DynamoDb. Only used in defining new DynVal instances.

Constructors

DynString

Fields

unDynString :: Text

newtype DynBool

Boolean values stored in DynamoDb. Only used in defining new DynVal instances.

Constructors

DynBool

Fields

unDynBool :: Bool

Working with key/value pairs

data Attribute

A key-value pair

Constructors

Attribute

Fields

attrName :: Text

attrVal :: DValue

parseAttributeJson :: Value -> Parser [Attribute]

Parse a JSON object that contains attributes

attributeJson :: Attribute -> Pair

Convert into JSON pair

attributesJson :: [Attribute] -> Value

Convert into JSON object for AWS.

attrTuple :: Attribute -> (Text, DValue)

Convert attribute to a tuple representation

attr :: DynVal a => Text -> a -> Attribute

Convenience function for constructing key-value pairs

attrAs :: DynVal a => Proxy a -> Text -> a -> Attribute

attr with type witness to help with cases where you're manually supplying values in code.

> item [ attrAs text "name" "john" ]

text :: Proxy Text

Type witness for Text. See attrAs.

int :: Proxy Integer

Type witness for Integer. See attrAs.

double :: Proxy Double

Type witness for Double. See attrAs.

data PrimaryKey

Primary keys consist of either just a Hash key (mandatory) or a hash key and a range key (optional).

Constructors

PrimaryKey

Fields

pkHash :: Attribute

pkRange :: Maybe Attribute

hk :: Text -> DValue -> PrimaryKey

Construct a hash-only primary key.

>>> hk "user-id" "ABCD"

>>> hk "user-id" (mkVal 23)

hrk

Arguments

:: Text
Hash key name

-> DValue
Hash key value

-> Text
Range key name

-> DValue
Range key value

-> PrimaryKey

Construct a hash-and-range primary key.

Working with objects (attribute collections)

type Item = Map Text DValue

A DynamoDb object is simply a key-value dictionary.

item :: [Attribute] -> Item

Pack a list of attributes into an Item.

attributes :: Map Text DValue -> [Attribute]

Unpack an Item into a list of attributes.

class ToDynItem a where

Types convertible to DynamoDb Item collections.

Use attr and attrAs combinators to conveniently define instances.

Methods

toItem :: a -> Item

class FromDynItem a where

Types parseable from DynamoDb Item collections.

User getAttr family of functions to applicatively or monadically parse into your custom types.

Methods

parseItem :: Item -> Parser a

fromItem :: FromDynItem a => Item -> Either String a

Parse an Item into target type using the FromDynItem instance.

newtype Parser a

A continuation-based parser type.

Constructors

Parser

Fields

runParser :: forall f r. Failure f r -> Success a f r -> f r

getAttr

Arguments

:: forall a. (Typeable a, DynVal a)

=> Text
Attribute name

-> Item
Item from DynamoDb

-> Parser a

Convenience combinator for parsing fields from an Item returned by DynamoDb.

getAttr'

Arguments

:: forall a. DynVal a

=> Text
Attribute name

-> Item
Item from DynamoDb

-> Parser (Maybe a)

Parse attribute if it's present in the Item. Fail if attribute is present but conversion fails.

parseAttr

Arguments

:: FromDynItem a

=> Text
Attribute name

-> Item
Item from DynamoDb

-> Parser a

Combinator for parsing an attribute into a FromDynItem.

Common types used by operations

data Conditions

Conditions used by mutation operations (PutItem, UpdateItem, etc.). The default def instance is empty (no condition).

Constructors

Conditions CondMerge [Condition]

conditionsJson :: Text -> Conditions -> [Pair]

JSON encoding of conditions parameter in various contexts.

expectsJson :: Conditions -> [Pair]

data Condition

A condition used by mutation operations (PutItem, UpdateItem, etc.).

Constructors

Condition

Fields

condAttr :: Text

Attribute to use as the basis for this conditional

condOp :: CondOp

Operation on the selected attribute

conditionJson :: Condition -> Pair

data CondOp

Conditional operation to perform on a field.

Constructors

DEq DValue

NotEq DValue

DLE DValue

DLT DValue

DGE DValue

DGT DValue

NotNull

IsNull

Contains DValue

NotContains DValue

Begins DValue

In [DValue]

Between DValue DValue

data CondMerge

How to merge multiple conditions.

Constructors

CondAnd

CondOr

data ConsumedCapacity

The standard response metrics on capacity consumption.

Constructors

ConsumedCapacity

Fields

capacityUnits :: Int64

capacityGlobalIndex :: [(Text, Int64)]

capacityLocalIndex :: [(Text, Int64)]

capacityTableUnits :: Maybe Int64

capacityTable :: Text

data ReturnConsumption

Constructors

RCIndexes

RCTotal

RCNone

data ItemCollectionMetrics

Constructors

ItemCollectionMetrics

Fields

icmKey :: (Text, DValue)

icmEstimate :: [Double]

data ReturnItemCollectionMetrics

Constructors

RICMSize

RICMNone

data UpdateReturn

What to return from the current update operation

Constructors

URNone
Return nothing

URAllOld
Return old values

URUpdatedOld
Return old values with a newer replacement

URAllNew
Return new values

URUpdatedNew
Return new values that were replacements

data QuerySelect

What to return from a Query or Scan query.

Constructors

SelectSpecific [Text]
Only return selected attributes

SelectCount
Return counts instead of attributes

SelectProjected
Return index-projected attributes

SelectAll
Default. Return everything.

querySelectJson :: KeyValue t => QuerySelect -> [t]

Size estimation

class DynSize a where

A class to help predict DynamoDb size of values, attributes and entire items. The result is given in number of bytes.

Methods

dynSize :: a -> Int

nullAttr :: Attribute -> Bool

Will an attribute be considered empty by DynamoDb?

A PutItem (or similar) with empty attributes will be rejected with a ValidationException.

Responses & Errors

data DdbResponse

Response metadata that is present in every DynamoDB response.

Constructors

DdbResponse

Fields

ddbrCrc :: Maybe Text

ddbrMsgId :: Maybe Text

data DdbErrCode

Errors defined by AWS.

Constructors

AccessDeniedException

ConditionalCheckFailedException

IncompleteSignatureException

InvalidSignatureException

LimitExceededException

MissingAuthenticationTokenException

ProvisionedThroughputExceededException

ResourceInUseException

ResourceNotFoundException

ThrottlingException

ValidationException

RequestTooLarge

InternalFailure

InternalServerError

ServiceUnavailableException

SerializationException
Raised by AWS when the request JSON is missing fields or is somehow malformed.

shouldRetry :: DdbErrCode -> Bool

Whether the action should be retried based on the received error.

data DdbError

Potential errors raised by DynamoDB

Constructors

DdbError

Fields

ddbStatusCode :: Int

200 if successful, 400 for client errors and 500 for server-side errors.

ddbErrCode :: DdbErrCode

ddbErrMsg :: Text

Internal Helpers

ddbSignQuery :: ToJSON a => ByteString -> a -> DdbConfiguration qt -> SignatureData -> SignedQuery

data AmazonError

Constructors

AmazonError

Fields

aeType :: Text

aeMessage :: Maybe Text

ddbResponseConsumer :: FromJSON a => IORef DdbResponse -> HTTPResponseConsumer a

ddbHttp :: Region -> DdbConfiguration NormalQuery

ddbHttps :: Region -> DdbConfiguration NormalQuery