Aws.Iam.Commands.CreateAccessKey

Documentation

data CreateAccessKey

Creates a new AWS secret access key and corresponding AWS access key ID for the given user name.

If a user name is not provided, IAM will determine the user name based on the access key signing the request.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_CreateAccessKey.html

Constructors

CreateAccessKey (Maybe Text)

data CreateAccessKeyResponse

Constructors

CreateAccessKeyResponse AccessKey

data AccessKey

Represents the IAM AccessKey data type.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_AccessKey.html

Constructors

AccessKey

Fields

akAccessKeyId :: Text

The Access Key ID.

akCreateDate :: Maybe UTCTime

Date and time at which the access key was created.

akSecretAccessKey :: Text

Secret key used to sign requests. The secret key is accessible only during key creation.

akStatus :: AccessKeyStatus

Whether the access key is active or not.

akUserName :: Text

The user name for which this key is defined.