Aws.Iam.Commands.DeleteGroupPolicy

Documentation

data DeleteGroupPolicy

Deletes the specified policy associated with the specified group.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_DeleteGroupPolicy.html

Constructors

DeleteGroupPolicy

Fields

dgpPolicyName :: Text

Name of the policy to be deleted.

dgpGroupName :: Text

Name of the group with whom the policy is associated.

data DeleteGroupPolicyResponse

Constructors

DeleteGroupPolicyResponse