Aws.Iam.Commands.GetUserPolicy

Documentation

data GetUserPolicy

Retreives the specified policy document for the specified user.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_GetUserPolicy.html

Constructors

GetUserPolicy

Fields

gupPolicyName :: Text

Name of the policy.

gupUserName :: Text

Name of the user with whom the policy is associated.

data GetUserPolicyResponse

Constructors

GetUserPolicyResponse

Fields

guprPolicyDocument :: Text

The policy document.

guprPolicyName :: Text

Name of the policy.

guprUserName :: Text

Name of the user with whom the policy is associated.