Aws.Iam.Commands.ListGroupPolicies

Documentation

data ListGroupPolicies

Lists the group policies associated with the specified group.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_ListGroupPolicies.html

Constructors

ListGroupPolicies

Fields

lgpGroupName :: Text

Policies associated with this group will be listed.

lgpMarker :: Maybe Text

Used for paginating requests. Marks the position of the last request.

lgpMaxItems :: Maybe Integer

Used for paginating requests. Specifies the maximum number of items to return in the response. Defaults to 100.

data ListGroupPoliciesResponse

Constructors

ListGroupPoliciesResponse

Fields

lgprPolicyNames :: [Text]

List of policy names.

lgprIsTruncated :: Bool

True if the request was truncated because of too many items.

lgprMarker :: Maybe Text

Marks the position at which the request was truncated. This value must be passed with the next request to continue listing from the last position.