Aws.Iam.Commands.ListGroups

Documentation

data ListGroups

Lists groups that have the specified path prefix.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_ListGroups.html

Constructors

ListGroups

Fields

lgPathPrefix :: Maybe Text

Groups defined under this path will be listed. If omitted, defaults to /, which lists all groups.

lgMarker :: Maybe Text

Used for paginating requests. Marks the position of the last request.

lgMaxItems :: Maybe Integer

Used for paginating requests. Specifies the maximum number of items to return in the response. Defaults to 100.

data ListGroupsResponse

Constructors

ListGroupsResponse

Fields

lgrGroups :: [Group]

List of Groups.

lgrIsTruncated :: Bool

True if the request was truncated because of too many items.

lgrMarker :: Maybe Text

Marks the position at which the request was truncated. This value must be passed with the next request to continue listing from the last position.

data Group

The IAM Group data type.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_Group.html

Constructors

Group

Fields

groupArn :: Text

ARN used to refer to this group.

groupCreateDate :: UTCTime

Date and time at which the group was created.

groupPath :: Text

Path under which the group was created.

groupGroupId :: Text

Unique identifier used to refer to this group.

groupGroupName :: Text

Name of the group.