Aws.Iam.Commands.ListMfaDevices

Documentation

data ListMfaDevices

Lists the MFA devices. If the request includes the user name, then this action lists all the MFA devices associated with the specified user name. If you do not specify a user name, IAM determines the user name implicitly based on the AWS access key ID signing the request.

https://docs.aws.amazon.com/IAM/latest/APIReference/API_ListMFADevices.html

Constructors

ListMfaDevices

Fields

lmfaUserName :: Maybe Text

The name of the user whose MFA devices you want to list. If you do not specify a user name, IAM determines the user name implicitly based on the AWS access key ID signing the request

lmfaMarker :: Maybe Text

Used for paginating requests. Marks the position of the last request.

lmfaMaxItems :: Maybe Integer

Used for paginating requests. Specifies the maximum number of items to return in the response. Defaults to 100.

data ListMfaDevicesResponse

Constructors

ListMfaDevicesResponse

Fields

lmfarMfaDevices :: [MfaDevice]

List of 'MFA Device's.

lmfarIsTruncated :: Bool

True if the request was truncated because of too many items.

lmfarMarker :: Maybe Text

Marks the position at which the request was truncated. This value must be passed with the next request to continue listing from the last position.