Aws.Iam.Core

Documentation

iamSignQuery

Arguments

:: [(ByteString, ByteString)]
Pairs of parameter names and values that will be passed as part of the request data.

-> IamConfiguration qt

-> SignatureData

-> SignedQuery

Constructs a SignedQuery with the specified request parameters.

iamResponseConsumer :: (Cursor -> Response IamMetadata a) -> IORef IamMetadata -> HTTPResponseConsumer a

Reads the metadata from an IAM response and delegates parsing the rest of the data from the response to the given function.

data IamMetadata

Constructors

IamMetadata

Fields

requestId :: Maybe Text

data IamConfiguration qt

Constructors

IamConfiguration

Fields

iamEndpoint :: ByteString

iamPort :: Int

iamProtocol :: Protocol

iamHttpMethod :: Method

data IamError

Constructors

IamError

Fields

iamStatusCode :: Status

iamErrorCode :: Text

iamErrorMessage :: Text

parseDateTime :: MonadThrow m => String -> m UTCTime

Parses IAM DateTime data type.

data AccessKeyStatus

Constructors

AccessKeyActive

AccessKeyInactive

data User

The IAM User data type.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_User.html

Constructors

User

Fields

userArn :: Text

ARN used to refer to this user.

userCreateDate :: UTCTime

Date and time at which the user was created.

userPath :: Text

Path under which the user was created.

userUserId :: Text

Unique identifier used to refer to this user.

userUserName :: Text

Name of the user.

parseUser :: MonadThrow m => Cursor -> m User

Parses the IAM User data type.

data Group

The IAM Group data type.

http://docs.aws.amazon.com/IAM/latest/APIReference/API_Group.html

Constructors

Group

Fields

groupArn :: Text

ARN used to refer to this group.

groupCreateDate :: UTCTime

Date and time at which the group was created.

groupPath :: Text

Path under which the group was created.

groupGroupId :: Text

Unique identifier used to refer to this group.

groupGroupName :: Text

Name of the group.

parseGroup :: MonadThrow m => Cursor -> m Group

Parses the IAM Group data type.

data MfaDevice

The IAM MFADevice data type.

https://docs.aws.amazon.com/IAM/latest/APIReference/API_MFADevice.html

Constructors

MfaDevice

Fields

mfaEnableDate :: UTCTime

The date when the MFA device was enabled for the user.

mfaSerialNumber :: Text

The serial number that uniquely identifies the MFA device. For virtual MFA devices, the serial number is the device ARN.

mfaUserName :: Text

The user with whom the MFA device is associated. Minimum length of 1. Maximum length of 64.

parseMfaDevice :: MonadThrow m => Cursor -> m MfaDevice

Parses the IAM MFADevice data type.