Aws.S3.Commands.GetObject

Documentation

data GetObject

Constructors

GetObject

Fields

goBucket :: Bucket

goObjectName :: Object

goVersionId :: Maybe Text

goResponseContentType :: Maybe Text

goResponseContentLanguage :: Maybe Text

goResponseExpires :: Maybe Text

goResponseCacheControl :: Maybe Text

goResponseContentDisposition :: Maybe Text

goResponseContentEncoding :: Maybe Text

goResponseContentRange :: Maybe (Int, Int)

goIfMatch :: Maybe Text

Return the object only if its entity tag (ETag, which is an md5sum of the content) is the same as the one specified; otherwise, catch a StatusCodeException with a status of 412 precondition failed.

goIfNoneMatch :: Maybe Text

Return the object only if its entity tag (ETag, which is an md5sum of the content) is different from the one specified; otherwise, catch a StatusCodeException with a status of 304 not modified.

getObject :: Bucket -> Text -> GetObject

data GetObjectResponse

Constructors

GetObjectResponse

Fields

gorMetadata :: ObjectMetadata

gorResponse :: Response (ConduitM () ByteString (ResourceT IO) ())

data GetObjectMemoryResponse

Constructors

GetObjectMemoryResponse ObjectMetadata (Response ByteString)