Aws.S3.Commands.HeadObject

Documentation

data HeadObject

Constructors

HeadObject

Fields

hoBucket :: Bucket

hoObjectName :: Object

hoVersionId :: Maybe Text

hoIfMatch :: Maybe Text

Return the object only if its entity tag (ETag, which is an md5sum of the content) is the same as the one specified; otherwise, catch a StatusCodeException with a status of 412 precondition failed.

hoIfNoneMatch :: Maybe Text

Return the object only if its entity tag (ETag, which is an md5sum of the content) is different from the one specified; otherwise, catch a StatusCodeException with a status of 304 not modified.

headObject :: Bucket -> Text -> HeadObject

data HeadObjectResponse

Constructors

HeadObjectResponse

Fields

horMetadata :: Maybe ObjectMetadata

data HeadObjectMemoryResponse

Constructors

HeadObjectMemoryResponse (Maybe ObjectMetadata)