Aws.Ses.Commands.SetIdentityNotificationTopic

Documentation

data SetIdentityNotificationTopic

Change or remove the Amazon SNS notification topic to which notification of the given type are published.

Constructors

SetIdentityNotificationTopic

Fields

sntIdentity :: Text

The identity for which the SNS topic will be changed.

sntNotificationType :: NotificationType

The type of notifications that will be published to the topic.

sntSnsTopic :: Maybe Text

Just the ARN of the SNS topic or Nothing to unset the topic.

data SetIdentityNotificationTopicResponse

The response sent back by SES after the SetIdentityNotificationTopic command.

Constructors

SetIdentityNotificationTopicResponse

data NotificationType

Constructors

Bounce

Complaint