Aws.Sqs.Core

Documentation

type ErrorCode = Text

data SqsError

Constructors

SqsError

Fields

sqsStatusCode :: Status

sqsErrorCode :: ErrorCode

sqsErrorType :: Text

sqsErrorMessage :: Text

sqsErrorDetail :: Maybe Text

sqsErrorMetadata :: Maybe SqsMetadata

SqsXmlError

Fields

sqsXmlErrorMessage :: Text

sqsXmlErrorMetadata :: Maybe SqsMetadata

data SqsMetadata

Constructors

SqsMetadata

Fields

sqsMAmzId2 :: Maybe Text

sqsMRequestId :: Maybe Text

data SqsAuthorization

Constructors

SqsAuthorizationHeader

SqsAuthorizationQuery

data Endpoint

Constructors

Endpoint

Fields

endpointHost :: ByteString

endpointDefaultLocationConstraint :: LocationConstraint

endpointAllowedLocationConstraints :: [LocationConstraint]

data SqsConfiguration qt

Constructors

SqsConfiguration

Fields

sqsProtocol :: Protocol

sqsEndpoint :: Endpoint

sqsPort :: Int

sqsUseUri :: Bool

sqsDefaultExpiry :: NominalDiffTime

sqsEndpointUsClassic :: Endpoint

sqsEndpointUsWest :: Endpoint

sqsEndpointUsWest2 :: Endpoint

sqsEndpointEu :: Endpoint

sqsEndpointEuWest2 :: Endpoint

sqsEndpointApSouthEast :: Endpoint

sqsEndpointApSouthEast2 :: Endpoint

sqsEndpointApNorthEast :: Endpoint

sqs :: Protocol -> Endpoint -> Bool -> SqsConfiguration qt

data SqsQuery

Constructors

SqsQuery

Fields

sqsQueueName :: Maybe QueueName

sqsQuery :: Query

sqsSignQuery :: SqsQuery -> SqsConfiguration qt -> SignatureData -> SignedQuery

sqsResponseConsumer :: HTTPResponseConsumer a -> IORef SqsMetadata -> HTTPResponseConsumer a

sqsXmlResponseConsumer :: (Cursor -> Response SqsMetadata a) -> IORef SqsMetadata -> HTTPResponseConsumer a

sqsErrorResponseConsumer :: HTTPResponseConsumer a

data QueueName

Constructors

QueueName

Fields

qName :: Text

qAccountNumber :: Text

printQueueName :: QueueName -> Text

data QueueAttribute

Constructors

QueueAll

ApproximateNumberOfMessages

ApproximateNumberOfMessagesNotVisible

VisibilityTimeout

CreatedTimestamp

LastModifiedTimestamp

Policy

MaximumMessageSize

MessageRetentionPeriod

QueueArn

data MessageAttribute

Constructors

MessageAll
all values

SenderId
the AWS account number (or the IP address, if anonymous access is allowed) of the sender

SentTimestamp
the time when the message was sent (epoch time in milliseconds)

ApproximateReceiveCount
the number of times a message has been received but not deleted

ApproximateFirstReceiveTimestamp
the time when the message was first received (epoch time in milliseconds)

data SqsPermission

Constructors

PermissionAll

PermissionSendMessage

PermissionReceiveMessage

PermissionDeleteMessage

PermissionChangeMessageVisibility

PermissionGetQueueAttributes

parseQueueAttribute :: MonadThrow m => Text -> m QueueAttribute

printQueueAttribute :: QueueAttribute -> Text

parseMessageAttribute :: MonadThrow m => Text -> m MessageAttribute

printMessageAttribute :: MessageAttribute -> Text

printPermission :: SqsPermission -> Text

newtype ReceiptHandle

Constructors

ReceiptHandle Text

newtype MessageId

Constructors

MessageId Text

printReceiptHandle :: ReceiptHandle -> Text